feat: added global storage
